+++ LeagueManagementSystem-Backend/src/main/java/ca/leaguemanagementsystem/dao/leagues/requests/RequestRepository.java
@@ -3,5 +3,5 @@ package ca.leaguemanagementsystem.dao.leagues.requests;
 import ca.leaguemanagementsystem.model.leagues.requests.Request;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RequestRepository extends CrudRepository<Request, Integer> {
+public interface RequestRepository extends CrudRepository<Request, long> {
 }
