+++ LeagueManagementSystem-Backend/src/main/java/ca/leaguemanagementsystem/dao/notifications/NotificationRepository.java
@@ -3,5 +3,5 @@ package ca.leaguemanagementsystem.dao.notifications;
 import ca.leaguemanagementsystem.model.notifications.Notification;
 import org.springframework.data.repository.CrudRepository;
 
-public interface NotificationRepository extends CrudRepository<Notification, Integer> {
+public interface NotificationRepository extends CrudRepository<Notification, long> {
 }
